move executor schema file location
