+++ libs/lib_base/src/main/java/com/bael/kirin/lib/base/service/BaseService.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.lifecycleScope
 import com.bael.kirin.lib.logger.contract.Logger
 import com.bael.kirin.lib.threading.contract.Threading
 import com.bael.kirin.lib.threading.executor.Executor
-import com.bael.kirin.lib.threading.executor.schema.ExecutorSchema
+import com.bael.kirin.lib.threading.executor.ExecutorSchema
 import com.bael.kirin.lib.threading.util.Util.MainThread
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
