+++ libs/lib_network/src/main/java/com/bael/kirin/lib/network/interactor/BaseInteractor.kt
@@ -3,7 +3,7 @@ package com.bael.kirin.lib.network.interactor
 import com.bael.kirin.lib.logger.contract.Logger
 import com.bael.kirin.lib.threading.contract.Threading
 import com.bael.kirin.lib.threading.executor.Executor
-import com.bael.kirin.lib.threading.executor.schema.ExecutorSchema
+import com.bael.kirin.lib.threading.executor.ExecutorSchema
 import com.bael.kirin.lib.threading.util.Util.IOThread
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
