+++ libs/lib_threading/src/main/java/com/bael/kirin/lib/threading/executor/Executor.kt
@@ -1,7 +1,5 @@
 package com.bael.kirin.lib.threading.executor
 
-import com.bael.kirin.lib.threading.executor.schema.ExecutorSchema
-
 /**
  * Created by ErickSumargo on 01/06/20.
  */
